CSHW-5: writing new orders to db instead of redis

+++ Ozon.Route256.Practice/src/Ozon.Route256.Practice.OrdersService/Dal/Models/OrderDal.cs
@@ -10,7 +10,7 @@ public record OrderDal(
     int AddressId,
     int ItemsCount,
     decimal Price,
-    float Weight,
+    double Weight,
     DateTime Created,
     string RegionName);
